refactor(Core/Mail): Move server mail into it's own file
